Move datatemplate to its own file
