+++ FormsPlayground/FormsPlayground/Features/Home/Views/FeatureView.xaml
@@ -1,0 +1,47 @@
+<?xml version="1.0" encoding="utf-8"?>
+
+<xamForms:ReactiveViewCell
+    x:TypeArguments="model:FeatureNode" 
+    xmlns="http://xamarin.com/schemas/2014/forms"
+    xmlns:x="http://schemas.microsoft.com/winfx/2009/xaml"
+    xmlns:xamForms="clr-namespace:ReactiveUI.XamForms;assembly=ReactiveUI.XamForms"
+    xmlns:model="clr-namespace:FormsPlayground.Features.Home.Model;assembly=FormsPlayground"
+    x:Class="FormsPlayground.Features.Home.Views.FeatureView"
+    x:DataType="model:FeatureNode">
+    
+    <Grid 
+        VerticalOptions="CenterAndExpand" 
+        HeightRequest="80" 
+        Padding="0">
+        <Label
+            Text="❯"
+            Margin="-14,-10,0,0"
+            TextColor="LightGray"
+            FontSize="50"
+            HorizontalOptions="Start"
+            VerticalOptions="Center" />
+        <StackLayout 
+            Padding="15,0"
+            Spacing="3" 
+            VerticalOptions="CenterAndExpand">
+            <Label
+                Margin="10,12,10,0"
+                TextColor="#414040"
+                FontSize="14"
+                FontAttributes="Bold"
+                Text="{Binding Title}" />
+            <Label
+                Margin="10,0,28,12"
+                FontSize="11"
+                Text="{Binding Subtitle}" />
+        </StackLayout>
+        <Label
+            x:Name="RightChevron"
+            Text="❯"
+            Margin="0,0,7,0"
+            TextColor="#414040"
+            FontSize="13"
+            HorizontalOptions="End"
+            VerticalOptions="Center" />
+    </Grid>
+</xamForms:ReactiveViewCell>
