+++ FormsPlayground/FormsPlayground/FormsPlayground.csproj
@@ -30,6 +30,9 @@
   </ItemGroup>
 
   <ItemGroup>
+    <EmbeddedResource Update="Features\Home\Views\FeatureView.xaml">
+      <Generator>MSBuild:Compile</Generator>
+    </EmbeddedResource>
     <EmbeddedResource Update="Resources\Strings.resx">
       <Generator>PublicResXFileCodeGenerator</Generator>
       <LastGenOutput>Strings.Designer.cs</LastGenOutput>
@@ -43,6 +46,13 @@
   <ItemGroup>
     <EmbeddedResource Include="Features\Home\Data\HomeDesign.json" />
     <EmbeddedResource Include="Features\Home\Data\HomeForms.json" />
+  </ItemGroup>
+
+  <ItemGroup>
+    <Page Include="Features\Home\Views\FeatureView.xaml">
+      <SubType>Designer</SubType>
+      <Generator></Generator>
+    </Page>
   </ItemGroup>
